MNIST_webapp/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_tflow_select=2.3.0=mkl
absl-py=0.15.0=pyhd3eb1b0_0
appnope=0.1.2=py37hecd8cb5_1001
astor=0.8.1=py37hecd8cb5_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bottleneck=1.3.4=py37h67323c0_0
brotli=1.0.9=hb1e8313_2
c-ares=1.18.1=hca72f7f_0
ca-certificates=2022.4.26=hecd8cb5_0
cached-property=1.5.2=py_0
certifi=2022.5.18.1=py37hecd8cb5_0
charset-normalizer=2.0.12=pypi_0
click=8.1.3=pypi_0
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
debugpy=1.5.1=py37he9d5cce_0
decorator=5.1.1=pyhd3eb1b0_0
entrypoints=0.4=py37hecd8cb5_0
flask=2.1.2=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=hd8bbffd_0
gast=0.2.2=py37_0
giflib=5.2.1=haf1e3a3_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpcio=1.42.0=py37ha29bfda_0
h5py=2.10.0=pypi_0
hdf5=1.10.6=hdbbcd12_0
idna=3.3=pypi_0
importlib-metadata=4.11.3=py37hecd8cb5_0
intel-openmp=2021.4.0=hecd8cb5_3538
ipykernel=6.9.1=py37hecd8cb5_0
ipython=7.31.1=py37hecd8cb5_0
itsdangerous=2.1.2=pypi_0
jedi=0.18.1=py37hecd8cb5_1
jinja2=3.1.2=pypi_0
joblib=1.1.0=pypi_0
jpeg=9e=hca72f7f_0
jupyter_client=7.2.2=py37hecd8cb5_0
jupyter_core=4.10.0=py37hecd8cb5_0
keras=2.3.1=0
keras-applications=1.0.8=py_1
keras-base=2.3.1=py37_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.4.2=py37he9d5cce_0
lcms2=2.12=hf1fd2bf_0
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libpng=1.6.37=ha441bb4_0
libprotobuf=3.20.1=h8346a28_0
libsodium=1.0.18=h1de35cc_0
libtiff=4.2.0=hdb42f99_1
libwebp=1.2.2=h56c3ce4_0
libwebp-base=1.2.2=hca72f7f_0
lz4-c=1.9.3=h23ab428_1
markdown=3.3.4=py37hecd8cb5_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.1=py37hecd8cb5_1
matplotlib-base=3.5.1=py37hfb0c5b7_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py37h9ed2024_0
mkl_fft=1.3.1=py37h4ab4a9b_0
mkl_random=1.2.2=py37hb2f4e1b_0
munkres=1.1.4=py_0
ncurses=6.3=hca72f7f_2
nest-asyncio=1.5.5=py37hecd8cb5_0
numexpr=2.8.1=py37h2e5f0a9_0
numpy=1.21.6=pypi_0
openssl=1.1.1o=hca72f7f_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py37h743cdd8_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=py37hde71d04_0
pip=21.2.2=py37hecd8cb5_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.20.1=py37he9d5cce_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pydot=1.3.0=py37_0
pygments=2.11.2=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
python=3.7.13=hdfd78df_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2022.1=py37hecd8cb5_0
pyyaml=6.0=py37hca72f7f_1
pyzmq=22.3.0=py37he9d5cce_2
readline=8.1.2=hca72f7f_1
requests=2.28.0=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=py37h8c7af03_0
setuptools=61.2.0=py37hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
sqlite=3.38.3=h707629a_0
tensorboard=2.0.0=pyhb38c66f_1
tensorflow=2.0.0=mkl_py37hda344b4_0
tensorflow-base=2.0.0=mkl_py37h66b1bf0_0
tensorflow-estimator=2.0.0=pyh2649769_0
termcolor=1.1.0=py37hecd8cb5_1
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h5d9f67b_0
tornado=6.1=py37h9ed2024_0
tqdm=4.64.0=pypi_0
traitlets=5.1.1=pyhd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.9=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.13.3=py37hca72f7f_2
xz=5.2.5=hca72f7f_1
yaml=0.2.5=haf1e3a3_0
zeromq=4.3.4=h23ab428_0
zipp=3.8.0=py37hecd8cb5_0
zlib=1.2.12=h4dc903c_2
zstd=1.5.2=hcb37349_0
